CamleCase component names and replace to Elements
