+++ src/theme/index.js
@@ -1,5 +1,4 @@
 import React, { useMemo, useEffect } from 'react';
-import { Text, TextProps as TextPropsOriginal, Button } from 'rebass';
 import styled, {
 	css,
 	ThemeProvider as StyledComponentsThemeProvider,
@@ -195,138 +194,3 @@ export default function ({ children }) {
 	);
 }
 
-const TextWrapper = styled(Text)`
-	color: ${({ color, theme }) =>
-		color ? theme[color] : theme['textPrimaryColor']};
-	font-size: ${({ fontSize, theme }) => theme[fontSize]};
-	line-height: ${({ lineHeight, theme }) => theme[lineHeight]};
-	text-align: 'left';
-	font-weight: ${({ fontWeight, theme }) =>
-		fontWeight ? theme[fontWeight] : theme['regular']};
-	margin-left: ${(props) =>
-		props.timestamp &&
-		(props.theme.isMobile()
-			? props.theme.spacing(5.5)
-			: props.theme.spacing(12))} !important;
-	margin-bottom: ${(props) =>
-		props.timestamp && props.theme.spacing(1.5)} !important;
-	padding-right: ${(props) =>
-		props.pr ? props.theme.spacing(props.pr) : 0} !important;
-	padding-left: ${(props) =>
-		props.pl ? props.theme.spacing(props.pl) : 0} !important;
-	white-space: ${(props) => (props.nowrap ? 'nowrap' : 'wrap')} !important;
-`;
-
-export const TYPOGRAPHY = {
-	headLine1(props) {
-		return (
-			<TextWrapper
-				fontWeight='bold'
-				fontSize='textH1'
-				lineHeight='lineHeightH1'
-				{...props}
-			/>
-		);
-	},
-	headLine2(props) {
-		return (
-			<TextWrapper
-				fontWeight='bold'
-				fontSize='textH2'
-				lineHeight='lineHeightH2'
-				{...props}
-			/>
-		);
-	},
-	headLine3(props) {
-		return (
-			<TextWrapper
-				fontWeight='bold'
-				fontSize='textH3'
-				lineHeight='lineHeightH3'
-				{...props}
-			/>
-		);
-	},
-	headLine4(props) {
-		return (
-			<TextWrapper
-				fontWeight='bold'
-				fontSize='textH4'
-				lineHeight='lineHeightH4'
-				color={props.label && 'textLabelColor'}
-				{...props}
-			/>
-		);
-	},
-	headLine5(props) {
-		return (
-			<TextWrapper
-				fontWeight='bold'
-				fontSize='textH5'
-				lineHeight='lineHeightH5'
-				{...props}
-			/>
-		);
-	},
-	body1(props) {
-		return (
-			<TextWrapper
-				fontWeight={props.bold ? 'bold' : 'regular'}
-				fontSize='textBody1'
-				lineHeight='lineHeightBody1'
-				{...props}
-			/>
-		);
-	},
-	body2(props) {
-		return (
-			<TextWrapper
-				fontSize='textBody2'
-				lineHeight={props.zarelaDay ? 'lineHeightDay' : 'lineHeightBody2'}
-				color={
-					props.sidebarHighLight
-						? '#3d5c8a'
-						: props.zarelaDay
-						? 'white'
-						: props.color
-				}
-				fontWeight={props.zarelaDay ? 'bold' : props.fontWeight}
-				{...props}
-			/>
-		);
-	},
-	timestamp(props) {
-		return (
-			<TextWrapper
-				fontSize='textHint'
-				lineHeight='lineHeightTimeStamp'
-				color='textTimeStampColor'
-				timestamp={true}
-				{...props}
-			/>
-		);
-	},
-	hint(props) {
-		return (
-			<TextWrapper
-				fontSize='textHint'
-				lineHeight='lineHeightHint'
-				nowrap
-				color={props.active ? 'textTokenColor' : 'textPrimaryColor'}
-				{...props}
-			/>
-		);
-	},
-	tag(props) {
-		return (
-			<TextWrapper
-				fontSize='textTag'
-				lineHeight='lineHeightTag'
-				nowrap
-				color='textTagColor'
-				{...props}
-			/>
-		);
-	},
-};
